Polling changes
Some changes to the polling functions which tidies up the loop function. There's now a function to reset the snake position and the game detects if a wall has been hit, and the restart button actually works. I've just realised we need to detect if the snake head touches the snake body though, that could be fun to sort out....